backend/requirements.txt
Django==5.2.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
openpyxl==3.1.5
psycopg2-binary==2.9.9
django-filter==24.3
Pillow==10.4.0
python-dotenv==1.0.1
requests==2.32.3